Fix associating records with lengths in FIT files (#74)
Use the length's start time and duration to determine which records belong to
a length, instead of using the timestamp.  The timestamp is the time when the
message was written to the FIT file not the time when the actual event
happened.

This fixes an issue where the records in a FIT file would be associated with
the wrong length.  Note that lengths are only used for lap swimming activities
and are a sub-division of a lap.