Fix leaking SECTION_SUMMARY rows when deleting sessions
When sessions were deleted, SECTION_SUMMARY rows for laps and lengths were not
deleted, and these rows were not accessible and useless.  Fixed the code to
correctly delete the rows and also added tests for this.

Note that this defect DID NOT corrupt the data, it just left garbaged data
behind.  This is now fixed.