Use a default set of columns for large qresults-list% widgets
Some list view widgets, such as activities view and lap view have lots of
available columns -- chose a sensibe default for when no preferences are
stored.  The user can always use "View/Setup Columns" to choose what columns
are available.  This improves the look of the application for users who run it
for the first time, but does not affect other users.