Updated README, submodules and build
Simplified the README file, updated the submodules and fixed the build.  Linux
builds started hanging and timing out on Azure Pipelines and could not find
the cause (they also fail for older commits, which passed in the past).
Disabled the Linux build for now, and enabled to run the tests on Windows.