Assorted list of small fixes
* etc/scripts/azure-pipelines.yml (jobs): update to Racket 8.5

* rkt/view-activities.rkt (view-activities%
*activity-list-display-columns*): use #f for all invalid (N/A)
values, for consistency.

* rkt/gps-segments/view-gps-segments.rkt (view-gps-segments%
match-lv-columns): use #f for all invalid (N/A) values, for
consistency.

* rkt/models/sport-zone.rkt (get-zone-id-for-sport): if a
sport/sub-sport has no sport zone, try sport/#f instead, this
means, for example, that we use the "Cycling" sport zone for "Road
Cycling" activities.

* rkt/widgets/qresults-list.rkt (qresults-list%
export-data-as-csv): fixup exporter to quote values correctly and
to avoid printing an extra comma at the end of the line.