Update build process w.r.t installer
Build process was updated to generate a SHA256 manifest of all files in the
distribution, special care is taken to generate the manifest after the
executable is patched with the version number using `verpatch`.

This manifest is signed, but the signature file is external to the installer,
since the signing process executes after the installer.

Also added the CHANGES.md file to the installation as a README file and moved
some unused scripts to an "attic" folder.