Add AL2-Climb-Analysis, new icons and other small things.
Add the AL2-Climb-Analysis application to this repository and build and
package it for the Windows installer. Previous location here:
https://github.com/alex-hhh/AL2-Climb-Analysis

Added new icons for the ActivityLog2 and AL2-Climb-Analysis applications, plus
the file to generate the logos.

Added an updated, but experimental version of the `al2-profiler`

Added the SQLite version to the Help/About dialog, for diagostic purposes.