Add a "df-describe" option for activities
Available as the "Activity/Show session data frame summary..." menu item.

This shows a description of the data frame for the activity, and can be used
do determine what data series are present and how many have NA values -- this
is mostly a debugging tool for the data files.

Re-purposed the "SQL Export" dialog for this and also updated the dialog to
use a mono-spaced font and make it read only.