Handle correctly renaming activities in view-activities.rkt (AB#47)
When an activity is renamed such that it now matches (or no longer matches)
the text filter in the activity list, the filter update used to raise an
exception since the activity would be in the internal data set, but not in the
one shown to the user (which only contain activities which match the text
filter).

Updated the code to handle this without an exception being raised.