Fix bug related to reading settings in session inspector
When no preference file is present, the map widget in the session inspector
would attempt to read settings which did not exist, causing an exception to be
thrown -- this prevented viewing sessions in a new install of ActivityLog2.

Also added tests to check that the session inspector tabs can be opened on new
installations.