Use 0-150% range for Heart Rate, Power and Speed Reserve (AB#68)
... this makes plots look nicer, and the Heart Rate will line up with Power or
Speed Reserves (assuming HR and Power/Speed zones are properly set).

We use 150% for HR reserve, even though it can go up to 100% only.  This graph
will be lined up with the power and speed reserves and those can go above 100%
since they are relative to threshold power and speed, rather than maximum
ones.  This is a bit of a hack...