Assorted fixes to sport-charms.rkt and gps-segments.rkt
* rkt/sport-charms.rkt: fixed contracts to functions to use the
`positive?` function -- we don't allow zero for the values anyway.

* rkt/gps-segments/gps-segments.rkt: (find-nearby-sessions)
(find-nearby-sessions/last-import): allow GEOID level to be
specified as an argument, to control the size of the search area.
(good-segment-match?): changed pass criteria and allows the user
to pass an "average allowable distance" argument, which controls
how close the paths can be in order to be considered to be
similar.