Speed up findig similar routes + bugfix gps segments
When looking for similar routes, search for the middle point of the route, as
the start and end of the route is more likely to coincide with many different
routes (e.g. most routes start at home).

Also use `dtw/window` to find if two routes are similar -- this is much faster
than `dtw/memory-efficient`

Also relax the match criteria for finding GPS segments